day 29 password manager tkinter gui project completed
uses Tkinter Canvas to create logo image, with multiple entries for email, password, and website. Implemented a 'Generate Password' button which creates random password with a set number of letters, symbols, and numbers, and fills in the password entry. Also uses Pyperclip to copy password to clipboard after generating. When 'Add' button is pressed, a file 'data.txt' will be opened for Append, and formats and saves the details of the password to the file.