day 31 built tkinter ui flashcard app
App uses images and data files provided to create a flashcard ui. First showing the French word, after seconds, revealing the English translation. When the X button is pressed, shows a new card. When check button is pressed, that card is removed from the list of words and saved new list in a csv file to be read. When first time ran, app will read data from french_words.csv if words_to_learn.csv is not created yet.